+++ src/main/java/roomescape/presentation/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.presentation;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,7 +8,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import roomescape.controller.dto.response.ExceptionResponse;
+import roomescape.presentation.dto.response.ExceptionResponse;
 import roomescape.application.exception.ReservationNotFoundException;
 
 import java.util.stream.Collectors;
